chore: address changed namespace in mrk-te/ibexa-migration-bundle2:3.x but preserve compatibility with 2.x

+++ src/DependencyInjection/Compiler/ReplaceContentTypeServicePass.php
@@ -7,11 +7,13 @@ namespace vardumper\IbexaAutomaticMigrationsBundle\DependencyInjection\Compiler;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
-final class ReplaceContentTypeServicePass implements CompilerPassInterface
+inal class ReplaceContentTypeServicePass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container): void
     {
-        if (!class_exists('\\Ibexa\\Bundle\\Migration\\IbexaMigrationBundle') && !class_exists('\\Kaliop\\IbexaMigrationBundle\\IbexaMigrationBundle')) {
+        if (!class_exists('\\Ibexa\\Bundle\\Migration\\IbexaMigrationBundle')  /* ibexa/migrations */
+            && !class_exists('\\Kaliop\\IbexaMigrationBundle\\IbexaMigrationBundle') /* mrk-te/ibexa-migration-bundle2 backwards-compatibility with ^2.x */
+            && !class_exists('\\Kaliop\\IbexaMigrationBundle\\KaliopMigrationBundle')) { /* mrk-te/ibexa-migration-bundle current ^3.x */
             error_log('unable to activate IbexaAutomaticMigrationsBundle because neither ibexa/migrations nor kaliop/ibexa-migration-bundle is installed. Please install one of them.');
             return;
         }
